Speed up the tokio-based advance function by ~20%
I just re-wrote the advance function incrementally,
starting from a version with identical performance
as the message-passing advance function. Overhead
in the tokio version seems to be introduced by
spawning tasks into the runtime, and likely by
the clone().await constructs, where a spawned task
which is also shared is cloned and then awaited. The
performance gain here seems mainly to come from the
contents of the hash map being changed from a primitive
and a conserved to just a primitive.