Generalize remaining driver functions around Hydrodynamics
The code is now set up to run in isothermal mode
when hydro=iso is set (the default), or with
thermodynamics when hydro=euler. The latter
option will exit with a message saying the
Euler equations mode is still unfinished.

The architecture for including other hydro systems is
now mostly finished. Still to do:

- Port the generalized hydrodynamics to the
  task-based advance scheme.
- The type constraints on functions are overly
  complex due to additional traits to enable
  conversion of the Conserved type to an H5Type.
  This could be simplified by deriving H5Type in
  the hydro equations themselves. That in turn
  requires adding HDF5 as an opt-in feature.

To actually run with thermodynamics, we still need
to implement:

- Initial disk model with energy
- Viscous energy flux
- Intercell flux function
- Gravitational source terms for the energy
- The cooling prescription