Add an ItemizedSourceTerms struct
A side-effect of this commit is that the
Conserved trait now implies H5Type. It seems
to be a pattern that H5Type is sticking its
nose in all over the place. The alternative
was jumping through hoops to manually write
structs that were not H5Type. It's nifty that
H5Type structs can be composed, although we
are depending on h5py playing well with the
nested compound types.